requirements.txt
aiohttp==3.7.4.post0
albion-api-client==0.1.2
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
discord.py==1.7.2
greenlet==1.1.0
hurry.filesize==0.9
idna==2.10
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.9.1
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.21
style==1.1.0
typing-extensions==3.10.0.0
update==0.0.1
urllib3==1.26.4
yarl==1.6.3
